+++ scrapper/src/main/java/edu/java/clientService/ScrapperService.java
@@ -1,7 +1,7 @@
 package edu.java.clientService;
 
-import edu.java.dto.GenericResponse;
-import edu.java.dto.LinkUpdate;
+import edu.java.models.GenericResponse;
+import edu.java.models.LinkUpdate;
 
 public interface ScrapperService {
     GenericResponse<Void> updateLinks(LinkUpdate linkUpdate);
