+++ scrapper/src/test/java/edu/java/scrapper/api/httpClient/BotClientTest.java
@@ -3,8 +3,8 @@ package edu.java.scrapper.api.httpClient;
 import com.github.tomakehurst.wiremock.WireMockServer;
 import edu.java.api.dto.response.ApiErrorResponse;
 import edu.java.api.httpClient.BotClient;
-import edu.java.dto.GenericResponse;
-import edu.java.dto.LinkUpdate;
+import edu.java.models.GenericResponse;
+import edu.java.models.LinkUpdate;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
